requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
Pillow==8.1.1
pytz==2021.1
sqlparse==0.4.1
django-extensions==3.1.1
django-random-queryset==0.1.3
django-crispy-forms==1.11.1
phonenumbers==8.12.20
django-phonenumber-field==5.0.0
transliterate==1.10.2
django-tables2==2.3.4
djangorestframework==3.12.4

